backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
email-validator>=2.0.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles>=23.0.0
croniter>=1.3.0
apscheduler>=3.10.0
# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0